refactor: move isolate db wrapper and CallKeep log callback to common package
