+++ lib/common/fcm_isolate_database.dart
@@ -1,0 +1,13 @@
+import 'package:webtrit_phone/data/data.dart';
+
+class FCMIsolateDatabase extends AppDatabase {
+  FCMIsolateDatabase(super.e);
+
+  static FCMIsolateDatabase? _instance;
+
+  static FCMIsolateDatabase instance(executor) {
+    _instance ??= FCMIsolateDatabase(executor);
+
+    return _instance!;
+  }
+}
